+++ glfw/wl_init.c
@@ -28,6 +28,7 @@
 
 #define _GNU_SOURCE
 #include "internal.h"
+#include "wl_client_side_decorations.h"
 #include "backend_utils.h"
 #include "linux_desktop_settings.h"
 #include "../kitty/monotonic.h"
@@ -781,7 +782,7 @@ static const struct wl_registry_listener registryListener = {
 };
 
 
-GLFWAPI int glfwGetCurrentSystemColorTheme(void) {
+GLFWAPI GLFWColorScheme glfwGetCurrentSystemColorTheme(void) {
     return glfw_current_system_color_theme();
 }